remove reduce_logmeanexp too
